data-pipelines/dags/rds-app/requirements.txt
# RDS to Redshift requirements
apache-airflow==2.4.0
apache-airflow-providers-amazon==2.4.0
psycopg2-binary==2.9.3
boto3
requests==2.26.0
pyspark==3.1.2
kubernetes==21.7.0  # עדכון לגרסה קיימת
